refactor build dev
